pyproject.toml
[tool.poetry]
name = "drf-north-trans"
version = "0.1.0"
description = ""
authors = ["Juan-Camilo-Suarez <juancamilosuarez3@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.1.3"
djangorestframework = "3.14.0"
python-dotenv = "0.21.0"
psycopg2 = "2.9.5"
pillow = "^9.3.0"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]
black = "22.8.0"
click = "8.0.4"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
